chore: rename examples tool
